% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{create_game}
\alias{create_game}
\title{Set up a holoalphabetic game}
\usage{
create_game(
  game_letters = NULL,
  central = NULL,
  num_letters = 7,
  min_word_length = 4,
  dictionary = "normal",
  obscenities = FALSE
)
}
\arguments{
\item{game_letters}{User-selected letters to use in the game (optional). Will
result in a warning if more letters are chosen than \code{num_letters}. If
including, this should be a character string, e.g., "stb".}

\item{central}{The central (that is, required) letter (optional). If not
provided, will be randomly chosen from among the \code{game_letters}.}

\item{num_letters}{The number of letters for the game; defaults to 7, and
should be between 6 and 10.}

\item{min_word_length}{Expected length of words. Defaults to 4 letters, but
can be between 2 and 6.}

\item{dictionary}{Character string. Choice of how detailed of a dictionary to
use. Can be any of "slim", "broad", or "normal". Defaults to "normal",
which is recommended.}

\item{obscenities}{Logical. Should obscenities be included? Defaults to
FALSE.}
}
\value{
Regardless of whether you include any inputs, the function returns a
  game object \code{game} which can be used with \code{\link{play_game}}. The
  results of \code{create_game} should be assigned and used in
  \code{play_game}; running it on its own will have no effect.
}
\description{
\code{create_game} defines the terms of the game, choosing letters that
contain a pangram (i.e., all of them are used together in at least one
English word). The function is called automatically by
\code{\link{play_game}}, which then interactively plays the game. However,
you can intentionally use \code{create_game} to save a set of words.
}
\details{
Because of random sampling, it is possible that the internal function will
not find a set of letters with a pangram; if so, simply run the function
again.
}
\examples{
create_game()
create_game(num_letters = 6, game_letters = "s")
create_game(obscenities = TRUE)

}
\seealso{
\code{\link{play_game}} to play the game
}
