% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_finders.R
\name{has_pangram}
\alias{has_pangram}
\title{Check to see if a set of words has a pangram}
\usage{
has_pangram(game_letters, dictionary = normal)
}
\arguments{
\item{game_letters}{A character vector of letters in the format of
c("l", "e", "t", "r", "s") or "letrs"}

\item{dictionary}{A character list of words---defaults to the normal list
from the package.}
}
\value{
Logical: TRUE if a pangram exists; FALSE if not.
}
\description{
Given a list of letters and a dictionary, is there a pangram?
}
\examples{
has_pangram(c("l", "e", "t", "r", "s")) # yes
has_pangram(c("d", "r", "g", "t", "i")) # none
# or:
has_pangram("leters")

}
\seealso{
\code{\link{find_all_words}}
}
