% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holoalphabetic.R
\docType{package}
\name{holoalphabetic}
\alias{holoalphabetic}
\title{holoalphabetic: Interactive word games to find pangrams}
\description{
Like the New York Times' "Spelling Bee", the package gives you a set of 7 or
more letters and invites you to identify all possible words using those
letters. Find the pangram which uses all 7, or simply identify as many words
as possible with those letters. The word lists come from the Spell Checking
Oriented Word Lists (SCOWL) by Kevin Atkinson, which can be found at
\url{http://wordlist.aspell.net/scowl-readme}.
}
\section{holoalphabetic functions}{

The primary functions in holoalphabetic are \code{\link{play_game}}, which
lets you actually play the game, and \code{\link{create_game}}, which just
sets a game up for you. For more information, look at the main github site,
\url{https://github.com/jdbest/holoalphabetic/}
}

