% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{play_game}
\alias{play_game}
\title{Play the holoalphabetic game}
\usage{
play_game(game = NULL, keep_central_first = FALSE, restart = FALSE, ...)
}
\arguments{
\item{game}{A game object created by \code{\link{create_game}}. If null, the
default, \code{play_game} will create a new game object. Alternatively, if
using a previous game, \code{game} will contain both the game and
\code{state} variables.}

\item{keep_central_first}{Logical. Should the game always display the
"central" letter first when printing the letters. Defaults to FALSE.}

\item{restart}{Logical. Only useful when trying anew with an existing game.
Deletes previous guesses and starts with a clean slate, while using the
same letters (and therefore words).}

\item{...}{Arguments to pass to \code{\link{create_game}}, including
\code{num_letters}, \code{game_letters}, \code{min_word_length},
\code{dictionary}, and \code{obscenities}.}
}
\value{
Regardless of whether you include any inputs, the code will run a
  version of the game. If passed a game object from \code{create_game},
  the game will use the same letters and words. If passed a previously-played
  game from \code{play_game}, the game will pick up where you left off, and
  remind you of past guesses. Assign \code{play_game} to a variable to keep
  playing later.
}
\description{
\code{play_game} provides an interactive interface for guessing words
based on a game made by \code{\link{create_game}}. If you call this function
directly, it will create the game itself. Assigning the results to a variable
will allow you to pause and continue the game later by running
\code{play_game(prev_game)}.
}
\seealso{
\code{\link{create_game}} to set up a game without playing.
}
