% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hint_cheat.R
\name{print_answers}
\alias{print_answers}
\title{Print all answers for a set of letters}
\usage{
print_answers(
  game_letters = NULL,
  central = NULL,
  num_letters = 7,
  min_word_length = 4,
  dictionary = "normal",
  obscenities = FALSE,
  game = NULL,
  state = NULL
)
}
\arguments{
\item{game_letters}{The letters to identify words from.
These should be a character string (e.g., "polemic").
(Required, if \code{game} is not provided.)}

\item{central}{The central (that is, required) letter
(optional). If not provided, will be randomly chosen
from among the \code{game_letters}.}

\item{num_letters}{The number of letters for the game;
defaults to 7, and should be between 6 and 10.}

\item{min_word_length}{Expected length of words. Defaults
to 4 letters, but can be between 2 and 6.}

\item{dictionary}{Character string. Choice of how detailed
of a dictionary to use. Can be any of "slim", "broad",
or "normal". Defaults to "normal", which is recommended.}

\item{obscenities}{Logical. Should obscenities be included?
Defaults to FALSE.}

\item{game}{A \code{holoalphabetic} game from
\code{\link{play_game}} or \code{\link{create_game}}.
If included, no other parameters are required.}

\item{state}{State of a \code{holoalphabetic} game; provided
from current gameplay.}
}
\value{
When called directly (or through gameplay), this
  prints the answers to all words possible.
}
\description{
\code{print_answers} is a cheater function which identifies
all the possible answers for a set of given letters.
}
\details{
These words may or may not be the same as those identified
with other word sets; you can modify by changing the chosen
dictionary. This function is called by
\code{\link{play_game}} when choosing to get all answers
(with "[a]" entered instead of a guess).

This function will always check whether you \strong{really}
want to see the answers, requiring a "y" to continue.
}
\examples{
print_answers(game = create_game("craking"))
print_answers("balmntu")

}
\seealso{
\code{\link{play_game}} to play the game

\code{\link{print_hint_list}} to get hints
  rather than answers
}
