% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hint_cheat.R
\name{print_hint_list}
\alias{print_hint_list}
\title{Prints hints to help get all the answers}
\usage{
print_hint_list(
  game_letters = NULL,
  central = NULL,
  num_letters = 7,
  min_word_length = 4,
  dictionary = "normal",
  obscenities = FALSE,
  game = NULL,
  hint = NULL
)
}
\arguments{
\item{game_letters}{The letters to identify words from.
These should be a character string (e.g., "polemic").
(Required, if \code{game} is not provided.)}

\item{central}{The central (that is, required) letter
(optional). If not provided, will be randomly chosen
from among the \code{game_letters}.}

\item{num_letters}{The number of letters for the game;
defaults to 7, and should be between 6 and 10.}

\item{min_word_length}{Expected length of words. Defaults
to 4 letters, but can be between 2 and 6.}

\item{dictionary}{Character string. Choice of how detailed
of a dictionary to use. Can be any of "slim", "broad",
or "normal". Defaults to "normal", which is recommended.}

\item{obscenities}{Logical. Should obscenities be included?
Defaults to FALSE.}

\item{game}{A \code{holoalphabetic} game from
\code{\link{play_game}} or \code{\link{create_game}}.
If included, no other parameters are required.}

\item{state}{State of a \code{holoalphabetic} game; provided
from current gameplay.}
}
\value{
When called directly (or through gameplay), this
  prints hints to help you find all answers.
}
\description{
\code{print_hint_list} is a helper function which identifies
the \strong{number} of words for the given set of letters,
and (if desired) identifies what letters (two or three)
words begin with.
}
\details{
These words may or may not be the same as those identified
with other word sets; you can make minor changes
by changing the chosen
dictionary. This function is called by
\code{\link{play_game}} when choosing to get hints
(with "[h]" entered instead of a guess). You can choose
to get hints of the first two of three letters, or none.
}
\examples{
print_hint_list(game = create_game("craking"))
print_hint_list("balmntu")

}
\seealso{
\code{\link{play_game}} to play the game

\code{\link{print_answers}} to get all answers
}
