% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{play_game}
\alias{play_game}
\title{Play the holoalphabetic game}
\usage{
play_game(
  game = NULL,
  keep_central_first = FALSE,
  restart = FALSE,
  game_letters = NULL,
  central = NULL,
  num_letters = 7,
  min_word_length = 4,
  dictionary = "normal",
  obscenities = FALSE
)
}
\arguments{
\item{game}{A game object created by \code{\link{create_game}}. If null, the
default, \code{play_game} will create a new game object. Alternatively, if
using a previous game, \code{game} will contain both the game and
\code{state} variables.}

\item{keep_central_first}{Logical. Should the game always display the
"central" letter first when printing the letters. Defaults to FALSE.}

\item{restart}{Logical. Only useful when trying anew with an existing game.
  Deletes previous guesses and starts with a clean slate, while using the
  same letters (and therefore words).

The remaining arguments are passed to the \code{\link{create_game}} function
  and all can be left out.}

\item{game_letters}{User-selected letters to use in the game. Will
result in a warning if more letters are chosen than \code{num_letters}. If
including, this should be a character string, e.g., "stb".}

\item{central}{The central (that is, required) letter. If not
provided, will be randomly chosen from among the \code{game_letters}.}

\item{num_letters}{The number of letters for the game; defaults to 7, and
should be between 6 and 10.}

\item{min_word_length}{Expected length of words. Defaults to 4 letters, but
can be between 2 and 6.}

\item{dictionary}{Character string. Choice of how detailed of a dictionary to
use. Can be any of "slim", "broad", or "normal". Defaults to "normal",
which is recommended.}

\item{obscenities}{Logical. Should obscenities be included? Defaults to
FALSE.}
}
\value{
Regardless of whether you include any inputs, the code will run a
  version of the game. If passed a game object from \code{create_game},
  the game will use the same letters and words. If passed a previously-played
  game from \code{play_game}, the game will pick up where you left off, and
  remind you of past guesses. Assign \code{play_game} to a variable to keep
  playing later.
}
\description{
\code{play_game} provides an interactive interface for guessing words
based on a game made by \code{\link{create_game}}. If you call this function
directly, it will create the game itself. Assigning the results to a variable
will allow you to pause and continue the game later by running
\code{play_game(prev_game)}.
}
\examples{
\dontrun{
play_game()

# assuming a saved game called prev_game
play_game(game = prev_game)

# permit obscenities
play_game(keep_central_first = TRUE,
          obscenities = TRUE)
}

}
\seealso{
\code{\link{create_game}} to set up a game without playing.
}
